+++ docs-src/src/js/index.js
@@ -3,6 +3,11 @@ import "./search.js";
 
 window.Ulu = ulu;
 
+ulu.utils.fontAwesome.configureIcons();
+
+console.log(ulu.settings.getSettings());
+
+
 ulu.ui.page.init();
 ulu.ui.grid.init();
 ulu.ui.popover.init();