+++ js/utils/index.js
@@ -5,3 +5,4 @@ export * as id from "./id.js";
 export * as index from "./index.js";
 export * as pauseYoutubeVideo from "./pause-youtube-video.js";
 export * as fileSave from "./file-save.js";
+export * as fontAwesome from "./font-awesome.js";